nodes/tester/pyproject.toml
[project]
name = "tester"
version = "0.0.0"
authors = [{ name = "Your Name", email = "email@email.com" }]
description = "tester"
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "artefacts-cli>=0.9.3",
    "dora-rs >= 0.3.9",
    "posetree>=1.1.2",
    "transforms-py",
]

[dependency-groups]
dev = ["pytest >=8.1.1", "ruff >=0.9.1"]

[project.scripts]
tester = "tester.main:main"

[tool.ruff.lint]
extend-select = [
  "D",   # pydocstyle
  "UP"
]

[tool.uv.sources]
transforms-py = { git = "https://github.com/art-e-fact/transforms-py" }
